code/requirements.txt
numpy
pillow
torch
torchvision
matplotlib
scipy
tqdm
